src/test/javascript/spec/app/entities/ci-rule/ci-rule.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { CIRuleService } from 'app/entities/ci-rule/ci-rule.service';
import { ICIRule, CIRule } from 'app/shared/model/ci-rule.model';

describe('Service Tests', () => {
  describe('CIRule Service', () => {
    let injector: TestBed;
    let service: CIRuleService;
    let httpMock: HttpTestingController;
    let elemDefault: ICIRule;
    let expectedResult: ICIRule | ICIRule[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CIRuleService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new CIRule(0, 'AAAAAAA', false, false, false, currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            modifyDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CIRule', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            modifyDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            modifyDate: currentDate
          },
          returnedFromService
        );

        service.create(new CIRule()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CIRule', () => {
        const returnedFromService = Object.assign(
          {
            ruleName: 'BBBBBB',
            critical: true,
            useInReport: true,
            useInMicroservice: true,
            modifyDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            modifyDate: currentDate
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CIRule', () => {
        const returnedFromService = Object.assign(
          {
            ruleName: 'BBBBBB',
            critical: true,
            useInReport: true,
            useInMicroservice: true,
            modifyDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            modifyDate: currentDate
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CIRule', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
